src/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
courlan==0.9.5
dateparser==1.2.0
distro==1.9.0
frozenlist==1.4.1
googlesearch-python==1.2.3
greenlet==3.0.3
h11==0.14.0
htmldate==1.7.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
jusText==3.0.0
langcodes==3.3.0
lxml==4.9.4
multidict==6.0.4
openai==1.8.0
packaging==23.2
playwright==1.41.0
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyee==11.0.1
pytest==7.4.4
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tenacity==8.2.3
text-unidecode==1.3
tld==0.13
tqdm==4.66.1
trafilatura==1.6.4
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
yarl==1.9.4
